tlp-0.4: fix some (packaging) regressions
* src_install(): call emake with CONFFILE=... so that the config file gets
  installed properly

* Set TLP_ENABLE=0 in TLP's config file as it used to be prior to 0.4

This is a packaging issue - upstream's config file sets TLP_ENABLE=1, which
causes TLP to take effect immediately after installing it, whereas the
ebuild traditionally set it to 0. Since tlp-0.4, the upstream's
config file is edited and installed instead of a separate (Gentoo-only) file.
Forgot about editing TLP_ENABLE, which is fixed by this commit.